Remove examples which use the old opengl_cb API
I made an attempt to convert some of them, but in the end, meh, I can't
be bothered. Do it yourself if you care, PRs to restore any of these
example by using the render API might be accepted.

The qt_opengl example was converted before; just fixing some old
comments in it.

Fixes: #38